requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
pandas
numpy
scikit-learn
gunicorn==21.2.0
Werkzeug==2.3.7
